Rename PostActivate and PostPassivate listeners to clarify their use with a blocking consumer.
