+++ clustering/infinispan/embedded/spi/src/main/java/org/wildfly/clustering/infinispan/listener/PostPassivateBlockingListener.java
@@ -1,6 +1,6 @@
 /*
  * JBoss, Home of Professional Open Source.
- * Copyright 2021, Red Hat, Inc., and individual contributors
+ * Copyright 2022, Red Hat, Inc., and individual contributors
  * as indicated by the @author tags. See the copyright.txt file in the
  * distribution for a full listing of individual contributors.
  *
@@ -23,31 +23,32 @@
 package org.wildfly.clustering.infinispan.listener;
 
 import java.util.concurrent.CompletionStage;
-import java.util.function.BiConsumer;
-import java.util.function.Function;
+import java.util.function.Consumer;
 
 import org.infinispan.Cache;
+import org.infinispan.commons.util.concurrent.CompletableFutures;
 import org.infinispan.notifications.Listener;
-import org.infinispan.notifications.cachelistener.annotation.CacheEntryActivated;
-import org.infinispan.notifications.cachelistener.event.CacheEntryActivatedEvent;
+import org.infinispan.notifications.cachelistener.annotation.CacheEntryPassivated;
 import org.infinispan.notifications.cachelistener.event.CacheEntryEvent;
+import org.infinispan.notifications.cachelistener.event.CacheEntryPassivatedEvent;
 
 /**
- * Generic non-blocking post-activation listener that delegates to a blocking consumer.
+ * Generic non-blocking post-passivation listener that delegates to a blocking consumer.
  * @author Paul Ferraro
  */
 @Listener(observation = Listener.Observation.POST)
-public class PostActivateListener<K, V> extends CacheEventListenerRegistrar<K, V> {
+public class PostPassivateBlockingListener<K, V> extends CacheEventListenerRegistrar<K, V> {
 
-    private final Function<CacheEntryEvent<K, V>, CompletionStage<Void>> listener;
+    private Consumer<CacheEntryEvent<K, V>> listener;
 
-    public PostActivateListener(Cache<K, V> cache, BiConsumer<K, V> consumer) {
+    public PostPassivateBlockingListener(Cache<K, V> cache, Consumer<K> listener) {
         super(cache);
-        this.listener = new BlockingCacheEventListener<>(cache, consumer);
+        this.listener = new BlockingCacheEventListener<>(cache, listener);
     }
 
-    @CacheEntryActivated
-    public CompletionStage<Void> postActivate(CacheEntryActivatedEvent<K, V> event) {
-        return this.listener.apply(event);
+    @CacheEntryPassivated
+    public CompletionStage<Void> postPassivate(CacheEntryPassivatedEvent<K, V> event) {
+        this.listener.accept(event);
+        return CompletableFutures.completedNull();
     }
 }
